chore: keep network dependency for docs and future use
