+++ crates/storage/provider/src/traits/mod.rs
@@ -7,7 +7,7 @@ mod block;
 pub use block::*;
 
 mod header_sync_gap;
-pub use header_sync_gap::{HeaderSyncGap, HeaderSyncGapProvider};
+pub use header_sync_gap::HeaderSyncGapProvider;
 
 mod state;
 pub use state::StateWriter;
